CORE-1859 Fix mysql compatibility issues

+++ Bundles/Collector/tests/SprykerTest/Zed/Collector/Business/Exporter/Writer/AbstractTouchUpdaterTest.php
@@ -9,6 +9,7 @@ namespace SprykerTest\Zed\Collector\Business\Exporter\Writer;
 
 use Codeception\Test\Unit;
 use Propel\Runtime\Connection\ConnectionInterface;
+use Spryker\Shared\Propel\PropelConstants;
 use Spryker\Zed\Collector\Business\Exporter\Writer\Storage\TouchUpdaterSet;
 use Spryker\Zed\Collector\Business\Model\BulkTouchQueryBuilder;
 use Spryker\Zed\Collector\CollectorConfig;
@@ -63,10 +64,20 @@ class AbstractTouchUpdaterTest extends Unit
         $connection = $this->createConnectionMock();
 
         $touchUpdater = $this->createTouchUpdater();
+        $databaseEngine = $this->collectorConfig->getCurrentEngineName();
+        $expectedQuery = '';
 
-        $expectedQuery =
-            "UPDATE touchKeyTableName_value SET key = 'data_key1' WHERE touchKeyIdColumnName_value = 'new value'; \n"
-            . "UPDATE touchKeyTableName_value SET key = 'data_key2' WHERE touchKeyIdColumnName_value = 'new value2'";
+        if ($databaseEngine === $this->collectorConfig->getPostgresEngineName()) {
+            $expectedQuery =
+                "UPDATE touchKeyTableName_value SET key = 'data_key1' WHERE touchKeyIdColumnName_value = 'new value'; \n"
+                . "UPDATE touchKeyTableName_value SET key = 'data_key2' WHERE touchKeyIdColumnName_value = 'new value2'";
+        }
+
+        if ($databaseEngine === $this->collectorConfig->getMysqlEngineName()) {
+            $expectedQuery =
+                "UPDATE `touchKeyTableName_value` SET `key` = 'data_key1' WHERE `touchKeyTableName_value`.`touchKeyIdColumnName_value` = 'new value'; \n"
+                . "UPDATE `touchKeyTableName_value` SET `key` = 'data_key2' WHERE `touchKeyTableName_value`.`touchKeyIdColumnName_value` = 'new value2'";
+        }
 
         $connection->expects($this->once())
             ->method('exec')
@@ -87,8 +98,16 @@ class AbstractTouchUpdaterTest extends Unit
         $connection = $this->createConnectionMock();
 
         $touchUpdater = $this->createTouchUpdater();
+        $databaseEngine = $this->collectorConfig->getCurrentEngineName();
+        $expectedQuery = '';
 
-        $expectedQuery = 'DELETE FROM touchKeyTableName_value WHERE fk_touch IN (id_touch1,id_touch2)';
+        if ($databaseEngine === $this->collectorConfig->getPostgresEngineName()) {
+            $expectedQuery = 'DELETE FROM touchKeyTableName_value WHERE fk_touch IN (id_touch1,id_touch2)';
+        }
+
+        if ($databaseEngine ===  $this->collectorConfig->getMysqlEngineName()) {
+            $expectedQuery = 'DELETE FROM `touchKeyTableName_value` WHERE `fk_touch` IN (id_touch1,id_touch2)';
+        }
 
         $connection->expects($this->once())
             ->method('exec')
